Coreを独立したプロジェクトに分離
