+++ TRViS.LocalServers.BveTs/TimetableServerPlugin.cs
@@ -1,197 +1,39 @@
 using System;
-using System.Collections.Specialized;
-using System.IO;
 using System.Linq;
-using System.Net;
-using System.Net.Sockets;
 using System.Text;
-using System.Text.Json;
-using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 
 using AtsEx.PluginHost.Plugins;
 using AtsEx.PluginHost.Plugins.Extensions;
 
 using BveTypes.ClassWrappers;
 
-using TR.SimpleHttpServer;
-
 using TRViS.JsonModels;
+using TRViS.LocalServers.Core;
 
 namespace TRViS.LocalServers.BveTs;
 
 [Plugin(PluginType.Extension)]
-public partial class TimetableServerPlugin : PluginBase, IExtension
+public partial class TimetableServerPlugin : PluginBase, IExtension, ITimetableServerBridge
 {
-	const string LISTENER_PATH = "/";
-	const string JSON_FILE_MIME = "application/json";
-	const string TIMETABLE_FILE_MIME = JSON_FILE_MIME;
-	const string TIMETABLE_FILE_NAME = "timetable.json";
-	const string SYNC_SERVICE_PATH = "sync";
-	const string QR_HTML_FILE_NAME = "index.html";
-	const string SCENARIO_INFO_FILE_NAME = "scenario-info.json";
 	const string WORK_GROUP_ID = "1";
 	const string WORK_ID = "1-1";
 	const string TRAIN_ID = "1-1-1";
-	const int LISTENER_PORT = 58600;
-	const int PORT_RETRY_MAX = 10;
-	readonly IPAddress[] localAddressList;
-	readonly HttpServer server;
-	readonly int port;
-
-	readonly NameValueCollection additionalHeaders = [];
+	readonly TimetableServerCore server;
 
 	public TimetableServerPlugin(PluginBuilder builder) : base(builder)
 	{
-		additionalHeaders.Add("Access-Control-Allow-Origin", "*");
-		localAddressList = Dns.GetHostAddresses(Dns.GetHostName());
-		server = StartListener(out port);
+		server = new(this);
 
 		AddLaunchBrowserButtonToContextMenu();
 	}
 
-	HttpServer StartListener(out int port)
-	{
-		SocketException? _ex = null;
-
-		port = LISTENER_PORT;
-		for (int i = 0; i < PORT_RETRY_MAX; i++)
-		{
-			HttpServer? server = null;
-			try
-			{
-				server = new((ushort)port, httpHandler);
-				server.Start();
-				return server;
-			}
-			catch (SocketException ex)
-			{
-				server?.Dispose();
-				_ex = ex;
-				if (ex.SocketErrorCode == SocketError.AddressAlreadyInUse)
-				{
-					port = LISTENER_PORT + i + 1;
-					continue;
-				}
-				else
-					throw;
-			}
-		}
-
-		throw new InvalidOperationException("TimetableServerPlugin: Address already in use.", _ex);
-	}
-
-	private Task<HttpResponse> httpHandler(HttpRequest request)
-	{
-		try
-		{
-			return AcceptTcpClientAsync(request);
-		}
-		catch (Exception ex)
-		{
-			Console.WriteLine(ex);
-			return Task.FromResult(new HttpResponse(
-				status: "500 Internal Server Error",
-				ContentType: "text/plain",
-				additionalHeaders: additionalHeaders,
-				body: $"Internal Server Error: {ex.Message}\r\n{ex.StackTrace}"
-			));
-		}
-	}
-
-	static bool IsPathAcceptable(string path)
-		=> path switch
-		{
-			LISTENER_PATH or (LISTENER_PATH + QR_HTML_FILE_NAME) => true,
-			LISTENER_PATH + TIMETABLE_FILE_NAME => true,
-			LISTENER_PATH + SCENARIO_INFO_FILE_NAME => true,
-			LISTENER_PATH + SYNC_SERVICE_PATH => true,
-			_ => false
-		};
-
-	async Task<HttpResponse> AcceptTcpClientAsync(HttpRequest request)
-	{
-		string path = request.Path;
-		string pathWithoutQueryOrHash = path.Split('?', '#')[0];
-
-		if (!IsPathAcceptable(pathWithoutQueryOrHash))
-		{
-			return new HttpResponse(
-				status: "404 Not Found",
-				ContentType: "text/plain",
-				additionalHeaders: additionalHeaders,
-				body: "Not Found"
-			);
-		}
-
-		if (pathWithoutQueryOrHash is LISTENER_PATH or (LISTENER_PATH + QR_HTML_FILE_NAME))
-			return await GenResponseFromEmbeddedResourceAsync(QR_HTML_FILE_NAME, "text/html", additionalHeaders);
-		else if (pathWithoutQueryOrHash is LISTENER_PATH + SYNC_SERVICE_PATH)
-		{
-			return new(
-				status: "200 OK",
-				ContentType: JSON_FILE_MIME,
-				additionalHeaders: additionalHeaders,
-				body: GenerateSyncResponse()
-			);
-		}
-
-		if (!BveHacker.IsScenarioCreated)
-		{
-			return new(
-				status: "204 No Content",
-				ContentType: "text/plain",
-				additionalHeaders: additionalHeaders,
-				body: "No Content (Scenario Not Loaded)"
-			);
-		}
-
-		try
-		{
-			byte[] content = pathWithoutQueryOrHash switch
-			{
-				LISTENER_PATH + TIMETABLE_FILE_NAME => GenerateJson(),
-				LISTENER_PATH + SCENARIO_INFO_FILE_NAME => GenerateScenarioInfoJson(),
-				_ => throw new InvalidOperationException("TimetableServerPlugin: Invalid path.")
-			};
-
-			return new HttpResponse(
-				status: "200 OK",
-				ContentType: TIMETABLE_FILE_MIME,
-				additionalHeaders: additionalHeaders,
-				body: content
-			);
-		}
-		catch (Exception ex)
-		{
-			return new HttpResponse(
-				status: "500 Internal Server Error",
-				ContentType: "text/plain",
-				additionalHeaders: additionalHeaders,
-				body: $"Internal Server Error: {ex.Message}\r\n{ex.StackTrace}"
-			);
-		}
-	}
-
-	static readonly JsonSerializerOptions GenerateScenarioInfo_JsonSerializerOptions = new()
-	{
-		WriteIndented = false,
-		PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-	};
-	private byte[] GenerateScenarioInfoJson()
-	{
-		ResponseTypes.ScenarioInfo scenarioInfo = new(
+	public bool IsScenarioLoaded => BveHacker.IsScenarioCreated;
+	public ResponseTypes.ScenarioInfo? CurrentScenario => new(
 			routeName: BveHacker.ScenarioInfo.RouteTitle,
 			scenarioName: BveHacker.ScenarioInfo.Title,
 			carName: BveHacker.ScenarioInfo.VehicleTitle
 		);
-
-		return JsonSerializer.SerializeToUtf8Bytes(
-			scenarioInfo,
-			GenerateScenarioInfo_JsonSerializerOptions
-		);
-	}
 
 	static string RemoveSpaceCharBetweenEachChar(string str)
 	{
@@ -233,15 +75,13 @@ public partial class TimetableServerPlugin : PluginBase, IExtension
 		}
 		return sb.ToString();
 	}
-	static readonly JsonSerializerOptions GenerateJson_JsonSerializerOptions = new()
-	{
-		WriteIndented = false,
-		DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
-	};
 	private static readonly Regex beginWithSpeedLimitRegex = new(@"^([\d-]+)\s*\/(.+)$", RegexOptions.Compiled);
 	private static readonly Regex endWithSpeedLimitRegex = new(@"^(.+)\/\s*([\d-]+)$", RegexOptions.Compiled);
-	byte[] GenerateJson()
+	public WorkGroupData[]? GetWorkGroup()
 	{
+		if (!BveHacker.IsScenarioCreated)
+			return null;
+
 		Scenario scenario = BveHacker.Scenario;
 		ScenarioInfo scenarioInfo = BveHacker.ScenarioInfo;
 		TimeTable timeTable = scenario.TimeTable;
@@ -354,12 +194,16 @@ public partial class TimetableServerPlugin : PluginBase, IExtension
 			Works: [trvisWorkData]
 		);
 
-		return JsonSerializer.SerializeToUtf8Bytes<WorkGroupData[]>([trvisWorkGroupData], GenerateJson_JsonSerializerOptions);
+		return [trvisWorkGroupData];
 	}
 
-	byte[] GenerateSyncResponse()
+	public WorkGroupData[]? GetWorkGroupByWorkGroupId(string workGroupId) => TimetableServerBridgeUtils.GetWorkGroupByWorkGroupId(GetWorkGroup, workGroupId);
+	public WorkGroupData[]? GetWorkGroupByWorkId(string workId) => TimetableServerBridgeUtils.GetWorkGroupByWorkId(GetWorkGroup, workId);
+	public WorkGroupData[]? GetWorkGroupByTrainId(string trainId) => TimetableServerBridgeUtils.GetWorkGroupByTrainId(GetWorkGroup, trainId);
+
+	public SyncedData GetSyncedData()
 	{
-		SyncedData syncedData = BveHacker.IsScenarioCreated
+		return BveHacker.IsScenarioCreated
 			? new(
 				Location_m: BveHacker.Scenario.LocationManager.Location,
 				Time_ms: BveHacker.Scenario.TimeManager.TimeMilliseconds,
@@ -370,30 +214,12 @@ public partial class TimetableServerPlugin : PluginBase, IExtension
 				Time_ms: null,
 				CanStart: false
 			);
-
-		return JsonSerializer.SerializeToUtf8Bytes(syncedData, GenerateJson_JsonSerializerOptions);
-	}
-
-	private static async Task<HttpResponse> GenResponseFromEmbeddedResourceAsync(string fileName, string contentType, NameValueCollection additionalHeaders)
-	{
-		using Stream stream = currentAssembly.GetManifestResourceStream($"TRViS.LocalServers.BveTs.{fileName}");
-		long length = stream.Length;
-		byte[] content = new byte[length];
-		await stream.ReadAsync(content, 0, (int)length);
-
-		return new HttpResponse(
-			status: "200 OK",
-			ContentType: contentType,
-			additionalHeaders: additionalHeaders,
-			body: content
-		);
 	}
 
 	public override TickResult Tick(TimeSpan elapsed) => new ExtensionTickResult();
 
 	public override void Dispose()
 	{
-		server.Stop();
 		server.Dispose();
 	}
 }
