+++ TRViS.LocalServers.Core/TRViS.LocalServers.Core.csproj
@@ -1,0 +1,29 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFrameworks>net48;netstandard2.0</TargetFrameworks>
+		<LangVersion>12</LangVersion>
+		<Nullable>enable</Nullable>
+		<Version>1.1.0</Version>
+		<Authors>Tetsu Otter</Authors>
+		<Company>Tech Otter</Company>
+		<Product>TRViS</Product>
+		<Copyright>Copyright 2024 Tetsu Otter</Copyright>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<PackageProjectUrl>https://github.com/TetsuOtter/TRViS.LocalServers</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/TetsuOtter/TRViS.LocalServers</RepositoryUrl>
+		<RepositoryType>git</RepositoryType>
+		<RestorePackagesWithLockFile>true</RestorePackagesWithLockFile>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<EmbeddedResource Include="../ConnectHelper/dist/index.html" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<PackageReference Include="TR.SimpleHttpServer" Version="1.0.0.1" />
+		<PackageReference Include="TRViS.JsonModels" Version="1.0.2" />
+		<PackageReference Include="System.Text.Json" Version="8.0.4" />
+	</ItemGroup>
+
+</Project>
