+++ TRViS.LocalServers.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TRViS.LocalServers.BveTs", "TRViS.LocalServers.BveTs\TRViS.LocalServers.BveTs.csproj", "{03E5704C-8F53-490D-B487-4DC7291F1997}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TRViS.LocalServers.Core", "TRViS.LocalServers.Core\TRViS.LocalServers.Core.csproj", "{96EA5ADB-A3F0-4547-B9DE-344FBF1BE0AC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{03E5704C-8F53-490D-B487-4DC7291F1997}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{03E5704C-8F53-490D-B487-4DC7291F1997}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{03E5704C-8F53-490D-B487-4DC7291F1997}.Release|Any CPU.Build.0 = Release|Any CPU
+		{96EA5ADB-A3F0-4547-B9DE-344FBF1BE0AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{96EA5ADB-A3F0-4547-B9DE-344FBF1BE0AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{96EA5ADB-A3F0-4547-B9DE-344FBF1BE0AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{96EA5ADB-A3F0-4547-B9DE-344FBF1BE0AC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
